# api/web/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bintime TT",
        "description": "HTTP JSON API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.localhost/api",
            "description": "Api Bintime TT"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get a list of available users, broken down by pages and sorted.",
                "operationId": "UserAllUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Page Number",
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "registeredAt",
                                "-registeredAt",
                                "firstName",
                                "-firstName",
                                "lastName",
                                "-lastName",
                                "email",
                                "-email"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\OpenApi\\User\\UserPaginateSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Create a new user by email.",
                "operationId": "UserJoinByEmail",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Forms\\User\\User\\JoinByEmailForm"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\User\\User\\UserResponder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\ViolationResponder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\NotFoundResponder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\EntityResponder"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "description": "Update user by id.",
                "operationId": "UserEditProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Forms\\User\\User\\EditProfileForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\User\\User\\UserResponder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\ViolationResponder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\NotFoundResponder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\EntityResponder"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "description": "Delete user by id.",
                "operationId": "UserDeleteProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\NotFoundResponder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\EntityResponder"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}